Feat(Pieza): elimina el arhcivo del objeto valor pieza_caracterisitica, ya que se incluye en el mismo fichero de la clase pieza al no utilizarse. -Resuelve #11
